server/requirements.txt
pip
flask

pyserial
